python/requirements.txt
guizero
pyserial
iotc
